\name{scrime-internal}
\alias{buildSNPmat}
\alias{canberra2Mats}
\alias{check4Monomorphism}
\alias{checkATCG}
\alias{checkCatMat}
\alias{checkDist}
\alias{checkDistMat}
\alias{checkMatSNP}
\alias{checkX1X2}
\alias{chisqClass2}
\alias{chisqInd}
\alias{cohen2Mats}
\alias{compChisqPval}
\alias{compVarLevs}
\alias{constructMatCheck}
\alias{constructMatIA}
\alias{dist2Mats}
\alias{euclidean2Mats}
\alias{evalRedundancy}
\alias{getBinDomRec}
\alias{getBinGenotype}
\alias{getCall}
\alias{getDistance}
\alias{getListIdentity}
\alias{ia.samp}
\alias{manhattan2Mats}
\alias{mc2Mats}
\alias{maximum2Mats}
\alias{minkowski2Mats}
\alias{minrc}
\alias{minrc2Mats}
\alias{modeDist}
\alias{pcc2Mats}
\alias{print.simulatedSNPs}
\alias{print.simSNPglm}
\alias{recodeCount}
\alias{recodeFirst}
\alias{recodeNoHete}
\alias{removeRedundancy}
\alias{replaceNAs}
\alias{row2col}
\alias{sample.snpmat}
\alias{smc2Mats}
\alias{snp1norm2Mats}
\alias{weightMode}


\title{Internal scrime functions}
\description{Internal scrime functions.}
\details{
   These functions are not meant to be directly called by the user.
}
\author{Holger Schwender, \email{holger.schw@gmx.de}}
\keyword{internal}
