\name{showChanges}
\alias{showChanges}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Displaying Changes}
\description{
  Shows the content of the file CHANGES stored in a package.
}
\usage{
showChanges(pkg = "scrime")
}

\arguments{
  \item{pkg}{character string naming the package.}
}

\author{Holger Schwender, \email{holger.schwender@udo.edu}}

\keyword{documentation}
